% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmdata}
\alias{mmdata}
\title{Multiple-choice, Multiple-Attempt Data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 437 rows and 13 columns.
}
\usage{
mmdata
}
\description{
A real dataset from a calculus course at a public university where
they used an online learning platform to realize a multiple-attempt
format for multiple-choice test items. The resulting response matrix
has 437 examinees and 13 test items.
}
\keyword{datasets}
