% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirtmm.R
\name{score_scheme}
\alias{score_scheme}
\title{scoring using a scheme}
\usage{
score_scheme(X, scheme)
}
\arguments{
\item{scheme}{A scoring scheme vector}

\item{data}{A reponse matrix (the number of attempts)}
}
\description{
scoring using a scheme
}
