% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirtmm.R
\name{simsirt}
\alias{simsirt}
\title{Simulate SIRT-MM models}
\usage{
simsirt(N, M, K, thetas, b, a, c, g, d, MK = -1)
}
\arguments{
\item{N}{a number of subjects}

\item{M}{a number of items}

\item{K}{a number of answer options}

\item{thetas}{a N x 1 vector of thetas}

\item{b}{a M x 1 vector of b parameter}

\item{a}{a M x 1 vector of a parameter}

\item{g}{a M x V matrix of g parameter where V is the number of g parameters.}

\item{d}{a M x W matrix of d parameter where W is the number of d parameters.}

\item{MK}{a number of response categories (the number of max attempt + 1). -1 if answer-until-correct.}

\item{ct}{a M x 1 vector of c parameter}
}
\description{
Test
}
