\name{sirtmm-package}
\alias{sirtmm-package}
\alias{sirtmm}
\docType{package}
\title{
\packageTitle{sirtmm}
}
\description{
\packageDescription{sirtmm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sirtmm}
\packageIndices{sirtmm}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{sirtmm}

Maintainer: \packageMaintainer{sirtmm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
