% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirtmm.R
\name{estimate}
\alias{estimate}
\title{Person Estimation}
\usage{
estimate(mod, data = NULL, method = "EAP", max_value = 6)
}
\arguments{
\item{mod}{A SIRT-MM model}

\item{data}{A reponse matrix used to estimate thetas}

\item{method}{The estimation method (EAP, MLE, or WLE)}

\item{max_value}{The maximum absolute value for theta used for MLE estimation}
}
\description{
Person parameters
}
\examples{

# Replicating Lu et al. (2025) !! This version doesn't utilize regularization.
data(mmdata2)

# Remove four items that led to bad convergence
mmdata3 <- mmdata2[, -c(19, 16, 10, 13)]

# Maximum number of attempts set to 3
mmdata3[mmdata3 == 4] <- 3

# Fit two models
mod1 <- sirtmm(mmdata3,
    k = 4, mk = 3, itemtype = "SIRT-MM",
    ngs = 1, nds = 0, lambda1 = 0, lambda2 = 0,
    lambda3 = 0, max_nriter = 10, verbose = F, quadpts = 31
)
mod2 <- sirtmm(mmdata3,
    k = 4, mk = 3, itemtype = "SIRT-MM",
    ngs = 0, nds = 0, lambda1 = 0, lambda2 = 0,
    lambda3 = 0, max_nriter = 10, verbose = F, quadpts = 31
)

# Print item parameter estimates
print(mod1$itempar)
print(mod2$itempar)


# Print model fit statistics
print(anova(mod1, mod2))

# Estimate theta
mod1_est <- estimate(mod1)
mod2_est <- estimate(mod2)

# Compare theta estimates between the two models
plot(mod1_est$thetas, mod2_est$thetas)
}
